Remove table Oid for DML on partition table (#14578)
#2481 and greenplum-db/gporca#186 removed PartOid from Orca, but the OidCol
still exists. The unused DML column may cause an issue if there is a
redistribute motion on top of PartitionSelectorDML when we try to update on a
partition table.

The CPhysicalDML ComputeRequiredLocalColumns is still using that column.
CXformUpdate2DML forgot to remove the generation of the partOid column, but it
is not in the project list of CPartitionSelectorDML. When there is a
Redistribute motion between CPhysicalDML and CPartitionSelectorDML, the
RedistributeMotion will reference the PartOid column, so we end up with an
error when translating project list during DXL2PlStmt stage.

This patch fixes the issue by completely removing PartOid and all the related
stuff from Orca.