Set replication slot's restart_lsn to not greater than redo location
Previously we have faced multiple failures of incremental recovery due
to missing wal files for pg_rewind. Fixes have been made by commit
d0001b5 and 185c7d7. More context and discussions can be found in the
following PRs:

https://github.com/greenplum-db/gpdb/pull/11989
https://github.com/greenplum-db/gpdb/pull/11891
https://github.com/greenplum-db/gpdb/pull/14397
https://github.com/greenplum-db/gpdb/pull/14672

In summary, the problem has been that the upstream's WAL retention
logic (using physical replication slot) has failed to keep all the
necessary WAL files pg_rewind needs on the target segment.
Specifically, if wal_keep_segments GUC is not conservatively set,
upstream keeps WAL files up until the minimum restart_lsn among all
the physical replication slots.  However, in order for pg_rewind to
not fail fast due to missing WAL files on the target, we have to be
slightly more conservative and keep all the WAL files up until "the
checkpoint redo position BEFORE the minimum restart_lsn among all the
replication slots", because this is essentially the last common
checkpoint of the pg_rewind source and target segments. pg_rewind
(internally used by gprecoverseg incremental recovery) needs WAL files
up until this location to present on the target segment.

In order to fix the problem, our previous fixes heavily changed code
in KeepLogSeg() to figure out where is the last common checkpoint when
we can. The modified logic turned out difficult to understand yet
still missed some edge cases. So now instead of putting more patches,
we decided to take this opportunity to fix the problem with a
different approach.

We reverted the aforementioned two commits, and in this commit, we
instead use the replication slot's restart_lsn itself to remember the
last redo point location if the redo location is behind the received
lsn.  This way the divergence from upstream is much smaller and the
logic is easier to understand. More importantly, the replication
slot's restart_lsn stored on disk, even though may not be flushed
freshly right before a segment restart, it at least will keep more but
not less WALs than we need.

With the change in this commit, we essentially repurposed physical
replication slots to serve not only streaming replication, but also
incremental recovery (pg_rewind).

We also added test coverage for WAL removal/recycle for the first time.

Proposed-by: Ashwin Agrawal <aashwin@vmware>
Authored-by: Alexandra Wang <walexandra@vmware.com>
Reviewed-by: Ashwin Agrawal <ashwin@vmware>
(cherry picked from commit 055c57dd7017059eb527a9e40e311897fca2cbc1)

PS: The following changes were made for tests to be compatible with 6X_STABLE:
1. s/pg_switch_wal()/pg_switch_xlog()/
2. Added SELECT pg_sleep() after pg_ctl_start() to reduce flakiness

The following changes are specifically for Test 5 which we added in 7X for
additional test coverage for WAL retention in general:
3. Removed commands that set GUC wal_recycle because 6X_STABLE doesn't have it
and this test in 6X may not need it.  So far this test continues to succeed in
6X even without this GUC and I have not observed flakiness. One reason could be
that 6X caculates number of WALs to recycle differently from 7X. In the future
if test 5 becomes flaky, the lack of the wal_recycle GUC could be the potential
root cause and we should consider either backporting it or writing the test
differently.
4. s/pg_ls_waldir()/pg_ls_dir('./pg_xlog')/
5. Replaced pg_control_checkpoint() with a helper UDF pg_controldata_redo_lsn()
because 6X_STABLE does not have a built-in equivalent.