Fix binary swap test for pg_get_partition_def() fix
After pg_get_partition_def() was fixed to handle signed partition
bound values, the binary swap test started to fail because we now
surround some of the partition bound values with quotes. This is an
expected failure and we'll need to remove the tables and views before
running the binary swap test.

Expected failure diff:
```
+<   WHERE (((((l.relation IS NOT NULL) AND (l.database = d.oid)) AND (((c.relname <> 'gp_fault_strategy'::name) AND (c.relname <> 'locktest_master'::name)) OR (c.relname IS NULL))) AND (d.datname = current_database())) AND (l.gp_segment_id = (-1)))
+---
+>   WHERE (((((l.relation IS NOT NULL) AND (l.database = d.oid)) AND (((c.relname <> 'gp_fault_strategy'::name) AND (c.relname <> 'locktest_master'::name)) OR (c.relname IS NULL))) AND (d.datname = current_database())) AND (l.gp_segment_id = '-1'::integer))
```

Since many partitioned table DDLs are affected by this code change hence lots
of diffs in the binary swap test, added a helper script to figure out what
objects (TABLEs and VIEWs) need to be dropped. Also intentionally print the
DROP statements in CI output for more visibility of the temporary reduced test
coverage.