Fix pg_dump external partition logic for multi-level partition tables
The old logic was only able to handle simple partition tables that did
not contain subpartitions. Also, the query being used was a bit
excessive for the result produced. To fix the issue, we now do two
simple queries: one query that gets the external partitions of a
partition table and another query that gets the subroot hierarchy to
be used when dealing with a multi-level partition table. Using the two
query results and some result parsing, we now properly output the
ALTER TABLE EXCHANGE PARTITION to contain the required ALTER PARTITION
syntax for each subroot.

Fixes https://github.com/greenplum-db/gpdb/issues/5003.

Co-authored-by: Kalen Krempely <kkrempely@vmware.com>