using postgres query optimizer instead of ORCA to run resgroup group pipeline tests (#13990)
This is the backport of #13974.

The resource group pipeline uses ORCA as an optimizer by default. But as a resource management
tool, it's unimportant which optimizer we use.

So use postgres query optimizer instead of ORCA to run resource group pipeline tests. After that,
we can remove the file of resgroup_bypass_optimizer.source and resgroup_bypass_optimizer_1.source.