Fix incorrect result from hash join on char column
Issue:
The join of two columns, one varchar and one char, of matching values with
trailing spaces, generates empty output.

Example:
create table foo (a varchar(3)) distributed by (a);
create table bar (b char(3)) distributed by (b);
insert into foo values ('1 ');
insert into bar values ('1 ');
select a, b from foo join bar on a = b;

 a | p
---+---
(0 rows)

Root cause:
According to Postgres documentation, "Values of type character are physically
padded with spaces to the specified width n, and are stored and displayed that
way. However, trailing spaces are treated as semantically insignificant and
disregarded when comparing two values of type character."

This means, '1 ' of char(3) (1 trailing space) is stored as '1  ' (2 trailing
spaces). However, when it's used for comparison, such as in a join condition or
filter, all the trailing spaces are removed, so it's treated as '1'.

In our example, the join condition foo.a = bar.b doesn't specify casting.
Therefore, foo.a of varchar is cast to bpchar, and the join condition becomes
bar.b = (foo.a)::bpchar. The trailing spaces are removed for both foo.a and
bar.b. '1' == '1' matches, so the query should return 1 row.

However, ORCA generates an additional hash condition, where both foo.a and bar.p
are cast to text. The join condition becomes (bar.b = (foo.a)::bpchar) AND
((bar.p)::text = (foo.a)::text)). '1 ' of varchar stays '1 ' when cast to text.
However, '1  ' of char(3) becomes '1' when cast to text, cause casting is
considered comparison. '1 ' doesn't match '1', so 0 rows is returned.

The derivation of additional hash conditions is designed to optimize
multi-column join operations, and in itself has no fault. However, casting char
to text (ORCA's derived join condition) isn't equivalent to casting varchar to
char (user's join condition). This is because hash of char and hash of
varchar/text are two operator classes that belong to two different operator
families. If two opclasses belong to different opfamilies, the same operator
(equality in this case) invokes different internal functions. Casting foo.a to
char enforces a motion on foo, whereas casting bar.b to text enforces a motion
on bar.

Solution:
We fix this bug by addressing the opfamily mismatch. When deriving the
equivalence classes, instead of checking if the opfamilies of post-cast
types match, we check the matching of opfamilies of pre-cast types.

(cherry picked from commit 3dd1aeb976)