Fix minirepro utility to create view dependencies for explain queries (#14875)
Previously, when using minirepro on a query with explain command that accessed
materialized views/view , the object that the materialized view/view
used/accessed were not dumped with the minirepro. This was because while
rewriting queries in pg_rewrite_query, the utilities were not rewritten and
dumped into a result list.  Due to this in the current implementation while
getting the relids of queries containing explain command, we were getting the
relids of the materialized view/view and not of the used/accessed objects. With
this fix we are able to fetch the relids of both the materialized view/views
and the used objects. So the minirepro will be able to collect the objects DDL
used in the view.


Cherry-picked from 0c169d6ea685e80aadbb4f4e31f5c99ddf2a7b8c