Use partition key opclass to get equality operator for PARTITION BY LIST
When partitioning by list, the PARTITION BY transformation makes a
partition rule where each partition key uses an equality operator "="
that is not schema-qualified. Later on in the CREATE TABLE logic
during type coercion, it tries to look up the equality operator by
going through OprCacheHash which requires a hash key that contains the
search_path to look in. This works for general cases but not for when
the partition key is a custom type and its equality operator is in a
schema that is not in the session's search_path (nor default
pg_catalog). Things are even worse when dealing with pg_restore (and
in relation pg_upgrade) where the search_path is always set to blank
at the beginning of the SQL restore so even custom operators created
in the public schema would have issues.

To fix the issue, we simply need to fully-qualify the operator name
with its actual namespace name during PARTITION BY transformation so
that the OprCacheKey used later on will explicitly set the search_path
to the actual operator namespace.

Fixes https://github.com/greenplum-db/gpdb/issues/9405.