Run analyzedb after loading dump file in the pg_upgrade CI job
For the pg_upgrade CI job, we need to run ANALYZE on the catalog after
loading in the dump file so that the planner can generate good query
plans. Without the ANALYZE, the query plans will be generated with bad
statistics which can cause severe slowness in some of the pg_dump
catalog queries.