[6X Only] Use opclass to retrieve operators for partitioned table. (#14933)
[6X Only] Use opclass to retrieve operators for partitioned table.

Historically, we hardedcode searching path in Greenplum when retrieving
operators for partitioned table.  When inserting data into gpdb, it looks
up operators in the 'pg_catalog' schema.  That is to say, we cannot
partition a table with user defined types unless we create the type and
its associated operators inside the 'pg_catalog' schema.  This patch
helps resolve the problem by looking up operators using the opclass that we
use when defining the partitioned table.

Co-Authored-by: Zhenghua Lyu <kainwen@gmail.com>