Fix InterruptHoldoffCount not being reset issue
cases(#15278)

This commit addresses issue #15278, where the InterruptHoldoffCount
value was not always resetting to zero in certain abnormal
scenarios. This issue had the potential to impact the proper
functioning of the ProcessInterrupts process. The fix ensures that the
InterruptHoldoffCount value is properly reset to zero in all cases.

This issue is found in #15203 , we will add test cases in pr #15203

backport form https://github.com/greenplum-db/gpdb/pull/15279