Fix missing redistribute for CTAS/insert into randomly distributed table using Orca (#15363)
When inserting into a randomly distributed table, Orca adds a random
distribution motion to ensure tuples are properly distributed across the
segments. However, when this motion is a duplicate sensitive motion,
which we define as being on top of a replicated table or a universal
distributed (eg a generate_series function), we don't actually perform a
redistribute. Instead, we filter each segment by the segment id,
ensuring all tuples only exist on one segment. This is done for
correctness, and is the correct behavior.

In this case, the distribution spec was creating a strict random
distribution, and this will match the required strict random
distribution required by the CTAS/insert statement. While we had logic
to create a Random distribution spec if the motion we were adding was over a
universal/replicated spec in CDistributionSpecRandom::AppendEnforcers,
we were missing similar logic in
CDistributionSpecNonSingleton::AppendEnforcers.  Adding this logic
ensures we add a redistribute when necessary to randomly distribute tuples.

(cherry-picked from 2b18c6a)