gpstate -e: Fix gprecoverseg progress reporting
**Issue:**
Currently, `gpstate -e` may sometimes display incorrect or stalled progress for `gprecoverseg` when the main `gprecoverseg` process is terminated or interrupted (e.g. due to SSH disconnections).

**RCA:**
This happens because `gpstate -e` depends on `recovery_progress.file` to show progress to the user. However, when the main `gprecoverseg` process is terminated, this file is no longer updated and is not cleaned up, resulting in stalled or incorrect progress. A fix was attempted for a similar issue in greenplum-db#13412, but it required the user to manually remove the `gprecoverseg.lock` directory.

**Fix:**
This commit resolves the issue by checking whether the main `gprecoverseg` process is running or not (by reading the PID from `gprecoverseg.lock` and checking its status) before displaying progress. As a result, `gpstate` will not show any progress if `gprecoverseg` is not running.

**Testing:**
- Modified the behave test cases introduced in greenplum-db#13412, so that a dummy process is created whose `PID` is used to create a sample `gprecoverseg.lock` directory for testing.
- Added unit test cases for `is_gprecoverseg_running` function.