Fix "500 internal error" when gpfdist fails to receive data from network(6X_STABLE). (#15722)
There are retry mechanism on the gpdb segment side which is as client
to gpfdist server when writing into writable external tables. However,
the retry is not working because of gpfdist legacy code that will
close fstream when failed to receive data from socket.

This commit allow the client side retry sending post request.

The write_file_buffer is used by all POST request. The start buffer size will be 1M.
The maximum size will be 1G. We want all POST data fit in the write_file_buffer, so
that when network error happens, no data will be written into the fstream.
The sending buffer at segment size will be range from 32K to 128M, so 1G is large
enough to hold all the POST data.

If the Content-length is larger then 1G, an 400 will be returned.

Co-authored-by: Yongtao Huang <yongtaoh@vmware.com>