Do not preserve relfrozenxid of user tables for pg_dump --binary-upgrade
When running pg_dump with --binary-upgrade, the relfrozenxids are
preserved for each user table. However, there were many issues with
the logic:
1. For partition tables, there was no preservation SQL commands
outputted for subroot and child partitions. Only the root partition
had a SQL command be outputted to preserve its relfrozenxid.
2. For AO/CO tables, the relfrozenxid does not need to be preserved
since AO/CO tuples do not contain xid info.
3. After pg_upgrade runs pg_restore to recreate objects and update
relfrozenxids as per the dump file, pg_upgrade later bulk updates
pg_class to set the relfrozenxids to the source segment's datfrozenxid
for applicable pg_class tuples... which seemingly makes the
relfrozenxid preservation SQL commands in the dump file redundant.

Because of the above reasons, we disable the preservation of
relfrozenxid for user tables. This should help a bit in speeding up
upgrade and reducing the dump file size.