(6X) Refactor motion deadlock handler.
The idea that IC-UDP motion deadlock is related to join plan is not
accurate, the pattern of IC-UDP motion deadlock should be from the
whole plan's perspective, that is:

  An MPP slice's algorithm is
    1. fetch one tuple from motion
    2. fetch tuples from another motion

Because the step 1 fetching one tuple from motion is not a sync
operation, here by sync we mean each QE in this slice operates in the
same beat. Like fetching all tuples from a motion until finishing is a
sync operation, because each QE in the sending slice must have send
out End-Of-Stream to each QE in the receving slice, and there are no
more calls to this motion node.

Previous idea was the step 2 only from: inner plan of join and
SubPlans in Join's Qual or Join's joinqual. This is not
correct. SubPlans can come from any expression not related to join:
target entry, plan qual, joinqual, withcheck option...

SubPlans in Greenplum, if containing a motion, the motion has to be
under a material node to make the SubPlan rescannable. Previous
prefetch inner (for nestloop), prefetch qual, prefetch joinqual is all
based on this. Also note in 6X, HashSubPlan also will have a
material node over the motion.

This commit fixes the motion deadlock involving SubPlans. The fix
method is before actually executing the main plan, at the first time,
always pre-execute the material over motion node in the main plan
node's SubPlans. I also remove the old test `deadlock2.sql` which is
all covered in the new test file `motion_deadlock.sql`.

This is 6X version of main branch PR
https://github.com/greenplum-db/gpdb/pull/15842.

Fixes Issue: https://github.com/greenplum-db/gpdb/issues/15719

Refer to gpdb-dev mailing list discussion for more details:
https://groups.google.com/a/greenplum.org/g/gpdb-dev/c/Y4ajINeKeUw