Fix test failing on centos6
Step 'An entry to {action} {env} env var is added on all hosts of cluster'
is failing on centos6 platform. This step was added as part of https://github.com/greenplum-db/gpdb/pull/15504

Reason:
systemctl command is being used to restart sshd service which is not
present on centos6 so it is failing to restart sshd.

Fix:
Updated the step to check for the os platform and version. If it is a
centos os wth version 6 then use 'sudo service sshd restart' command to
restart sshd service.