Consider bitmap alternative only for ANY ScalarArray
Partial cherry-pick of commit 8bbadb87007502365b244f46d654d6ec3684ced3

Updated test for ALL to have single element in array so that ORCA goes
through the codepath that has been fixed as part of this commit. In GP6,
ORCA optimizes ALL with multiple array elements to a one-time false
filter.

Original commit messgae:

Commit 290166f introduced ScalarArrayOpExpr as an indexable qual for ANY
operator only as it is translated into an OR combination of indexscan on
the array elements.  Prior to this commit ORCA considered Bitmap
alternatives for equality on ScalarArrayCmp (both ANY and ALL) whereas
similar to planner, ORCA should also only consider it for equality on
ANY op. This commit fixes that issue and adds tests for the same.