Reorder scalar comparison expression in case of cast (#15880) (#15944)
Motivation:
Customer reports plan difference when the predicate is written in the form of
`col op val` vs. `val op col`. Specifically, if the column is the on left side,
ORCA is able to leverage the bitmap scan alternative given the indexed column.
When the value is placed on the left side, ORCA chooses much inferior a plan
with sequential scan.

Root cause:
The transform from select into bitmap scan expects predicates with columns on
the left side and values on the right side. That is, in order for the bitmap
scan alternative to be generated, the predicate has to be in the form of `col op
val`. If the predicate is written in the form `val op col`, preprocessor
converts it to `col op val`. This way, entering the exploration phase, a scalar
comparison expression is already reordered.

However, as of now, the expression reordering only works for `const op ident`,
but not for `const op cast(ident)`, which is the case of this query. Because of
this, the bitmap scan alternative isn't generated, and ORCA chooses the less
performant sequential scan.

Solution:
Support scalar comparison expression reordering in case of cast.

Implementation:
We implement this logic in the preprocessor instead of xform for two reasons. 1.
Preprocessor handles expression conversion recursively, so it works for nested
expressions. 2. Preprocessor has a broader impact than xform, and could
potentially benefit other xforms that expect predicates to be in the form of
`col op val`.

Test:
The scalar comparison expression order affects bitmap scan, but not index scan.
To demonstrate the fix allows the generation of the bitmap scan alternative, we
use 1. appendonly tables where index scan isn't supported, and 2. more than one
index when scanning a table.

Co-authored-by: Jingyu Wang <wjingyu@vmware.com>
(cherry picked from commit fe88dfc7657194fe8107f786f59f113ea6da5a80)

Fix https://github.com/greenplum-db/gpdb/issues/15624