Fix error of unique index key order (#16013)
The unique index key order will affect the distribution key order of child
partitions which causes an inconsistency issue in gpcheckcat. The root cause
is that transformDistributedBy() generated wrong distribution keys. The fix
is to avoid calling of transformDistributedBy() when adding a part to a
partition since distribution keys have been given at the time.