Changing dtx_phase2_retry_count to 60 (#16029)
In the previous code, the default value of dtx_phase2_retry_count is 10.

If the command cannot be dispatched successfully within this period, a PANIC happens (details are in doNotifyingCommitPrepared()). Then the postmaster will restart and goes into recovery process.

So, a small value may make user confused: why my postmaster restarts; but a big value is also not good: the Txn keeps retrying in dispatch, it may block other Txns.

After a long discussion: #15632, we choose a compromise default value: 60(near 10min) here.