[6X] Hintbit WAL syncrep should not throttle when holding exclusive lock
Backport from c5abb1f3e6c20f5fa372c2c807dfd22b350985f2

The only difference between the 6X and 7X change is that in 6X we can directly
use the "exclusive" field in the content lock. So we do not need to add a function
like 7X which was mainly for the purpose of not exposing the buffer exclusive
lock.

Commit description:

Similar to a previous deadlock issue (bd74bfc),
there's another deadlock possibility in hintbit WAL syncrep while mirror is down,
due to it is holding an exclusive buffer lock:

1. Mirror is down.
2. Primary does something that sets hintbits with exclusive lock on buffers,
   such as VACUUM.
3. During setting hintbits, we would throttle WAL rep (wait_to_avoid_large_repl_lag).
   Because mirror is down, this is essentially blocking until syncrep is disabled.
4. FTS detects mirror down, and tries to disable syncrep (setting GUC
   synchronous_standby_names to NULL). However, checkpointer as the only entity
   that really turns off syncrep is stuck flushing some buffer, because the VACUUM
   is holding exclusive lock on that buffer.
5. So, VACUUM is waiting for checkpointer to turn off syncrep but checkpointer
   is waiting for VACUUM to release lock. We have a deadlock situation.

We should never hold exclusive lock while doing syncrep. So fixing that. The
downside is that we will lose the throttling opportunity in that case, causing
WAL lag to build up. But that seems OK, because if we ever call
wait_to_avoid_large_repl_lag() again w/ a shared lock or w/o any lock at all,
we would still throttle. There are more places where we only acquire shared lock
or no lock than exclusive lock when running into wait_to_avoid_large_repl_lag().
For example, in the case of VACUUM, we have at least two such chances: when
scanning a tuple for the purpose of MVCC (shared lock is required), and when
finishing up scanning a block in lazy_scan_heap() (no lock is acquired).

Also renamed the "select_throttle" test to "hintbit_throttle" to better reflect
the reality.