change seq_segs update position (#16168)
Current gpfdist will update the corresponding sequence number when it responds a HTTP request successfully. However, the network may be unstable when gpfdist processes a HTTP request. In rare conditions, a connection could be aborted when gpfdist finishes receiving and writing data, which would lead to a data-sending retry.

Since gpfdist wouldn't update the sequence number despite finishing data writing when the response failed, gpdb will retry to send the same data which will be written again.

The PR is to handle the rare case. Gpfdist updates the sequence number when data writing ends or the request is OPEN_SEQ. This will avoid the unexpected network abortion when gpfdist responds 'OK' to gpdb. After the sequence number is updated for the segment, even if the same data is sent to gpfdist, The data will be discarded. The adaptation will avoid data overwriting.

If you want to reproduce the error, I suggest you do the following steps on gpfdist v6.24.5:

    Change the return value of http_ok to 1 when request_id in request_t is odd.
    Recompile the gpfdist and run it.
    Start any task that writes to a writable external table. Note that the data size should be bigger than 256MB.
    The actually written data is bigger the size of the expected value.

The core objective of the above steps is to simulate the specific network error in a hardcoded way.

We have tested the correctness of the modification by simulating the network error using the way of reproducing the error.