Archive current timeline history file after recovery finishes if needed
Previously, the only time the timeline history file would be archived
was when a standby already configured for WAL archiving was
promoted. If WAL archiving was set up and started after the standby
was promoted, its current timeline history file would not be found in
the archive. This could cause issues when restoring backups that did
not include the timeline history files (e.g. pg_basebackup
-Xnone). Some example failures include failing to restore with
recovery_target_timeline explicitly set to the control file's timeline
id and failing to create cascade standbys after recovering with
recovery_target_timeline set to 'current' or 'latest'. To prevent
these restore issues, we now ensure that the current timeline history
file has been archived by marking it as ready for archiving after
recovery finishes if it was found that it has not already been
archived or marked ready.

Backported from GPDB main:
https://github.com/greenplum-db/gpdb/commit/4e865bb2930df604387b566174061e9d6a26ac0e