Add support for pgbouncer to switch from ldap to md5 aunthentication. (#16286)
The PR of pgbouncer is https://github.com/greenplum-db/pgbouncer/pull/23.

Some customers have the need to switch authentication method
by changing the auth_hba_file and reload without restarting pgbouncer.
However the ldap implementation adds new users in a different user tree.
So when the customer switches the authentication method, it tries
to search a created pool on the new user->pools list. As the new user
and the old user have the same username, but do not point to the same
structure, so it will try to add a new pool to pool_list which
will trigger a fatal error because a pool with the same 'username' and
'dbname' has already been added in pool_list.

This commit will remove the difference between md5 and ldap user so
that the old and new user could share the same data structure. So there is
no need to add a new pool to pool_list when customer switch the
authentication method because the 'username' and 'dbname' have not changed.

The common steps is as follows:

1. Login with ldap authentication method
2. Modify hba.conf file to md5 method and reload
3. Touch userlist.txt(need to refresh user password which is important if last authentication method is ldap, so the password stored will be the ldap password, this step will refresh the password with the md5 password)
4. Relogin using the new authentication method like md5.