[6x]: Fix failed jobs for rocky9/oel9/rhel9 (#16264)
* Fix failed jobs for rocky9/oel9/rhel9

Support for 6x has been introduced for rocky9/oel9/rhel9, but there are some jobs experiencing failures.

Failure 1: gpcheckperf Failure
Root Cause - This failure is attributed to the absence of Python 2, as these operating systems exclusively
support Python 3. The code relied on Python 2-specific syntax, including the "print" statement.
Solution - To rectify this, we have diligently updated the affected code sections to be compatible with
Python 2 and Python 3 both. Instances of Python 2-specific syntax, such as the "print" statement, have
been replaced with Python 3-compatible alternatives like "print()".

Failure 2: gpinitsystem failure
Root Cause - When the banner is set through .bashrc this gpinitsystem failure was primarily caused by the presence of a banner in the command output,
making it challenging to identify and parse the necessary version information.
Solution - To overcome this issue, we have introduced a delimiter that effectively separates the banners from
the version information. This delimiter, 'START_CMD_OUTPUT\n', allows us to extract the version information by splitting the command's stdout.

Failure 3: unit_test_gpmgmt failure
Root Cause - The failure in unit_test occurred due to discrepancies between the expected and actual output.
Specifically, the error message format for invalid bash commands was inconsistent. The Bash version in Rocky 9 is 5.1.8, while in Rocky 8, it's 4.4.20. The difference in bash versions results in distinct error messages.
Solution - To address this issue, we have implemented a solution using a regular expression (regex) pattern.
This pattern accommodates varying formats of error messages for invalid bash commands, whether or not they include ": line".

Co-authored-by: Rakesh Sharma <shrakesh@vmware.com>