gpstop does not work if previous gpstop was interrupted
Issue:
When a gpstop process with 'smart' option is interrupted, the subsequent
gpstop command becomes stuck with an error message stating "the database
system is shutting down". This error message also appears when attempting
to connect to the database using psql and other utilities.
The problem is that the error message does not provide clear instructions
on how to stop the database, leading to confusion and an inability to
perform a graceful shutdown. Gpstop will not work at this point and the
only way to stop the database is to manually kill the master processes.

RCA
The issue occurs due to the interruption of the gpstop process, which
leaves the database system in an incomplete shutdown state with
unfinished master processes. Also, this will not allow any further
operations on the database until the shutdown is complete.

Solution:
This issue will be addressed by modifying the gpstop code to check if
the database is shutting down. If yes, it will show a warning stating
"The database is currently in the process of shutting down". After
this warning, the code will prompt the user to choose between a "Fast"
or "Immediate" shutdown option. Depending on the user's selection,
the code will proceed as follows:

- Terminate all master processes stuck in shutdown state from
previously interrupted gpstop.
- Start the cluster in master-only mode so that we can get the cluster
 information required in subsequent steps
- Continue the normal gpstop workflow.

Other changes include adding a 'pidlockpath' in mainOptions of gpstop.
This creates a gpstop.lock file under MASTER_DATA_DIRECTORY to
prevent the user from running multiple instances of gpstop.